network, tests: move the HTTP/TCP server creation to a separate file
