+++ tests/BUILD.bazel
@@ -10,6 +10,7 @@ go_library(
         "ping.go",
         "test.go",
         "utils.go",
+        "vmi_servers.go",
     ],
     importpath = "kubevirt.io/kubevirt/tests",
     visibility = ["//visibility:public"],
@@ -190,7 +191,6 @@ go_test(
         "//staging/src/kubevirt.io/client-go/subresources:go_default_library",
         "//tests/containerdisk:go_default_library",
         "//tests/flags:go_default_library",
-        "//tests/libvmi:go_default_library",
         "//tests/network:go_default_library",
         "//tests/reporter:go_default_library",
         "//tools/vms-generator/utils:go_default_library",
